add transactin entity

+++ app/src/main/java/com/fundtransfer/hcl/app/model/Account.java
@@ -25,7 +25,7 @@ public class Account {
 	private Date creationDate;
 	
 	@Column(name="BALANCE")
-	private Float balance;
+	private Double balance;
 
 	@ManyToOne
     private Customer customer;
@@ -62,11 +62,11 @@ public class Account {
 		this.creationDate = creationDate;
 	}
 
-	public Float getBalance() {
+	public Double getBalance() {
 		return balance;
 	}
 
-	public void setBalance(Float balance) {
+	public void setBalance(Double balance) {
 		this.balance = balance;
 	}
